+++ app/src/main/java/com/yourssu/unscramble/util/base/BindActivity.kt
@@ -1,10 +1,13 @@
-package com.yourssu.unscramble.base
+package com.yourssu.unscramble.util.base
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.viewbinding.ViewBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
 
 abstract class BindActivity<B : ViewBinding> : AppCompatActivity() {
