+++ app/src/main/java/com/yourssu/unscramble/util/extension/StringExtension.kt
@@ -1,4 +1,4 @@
-package com.yourssu.unscramble.util
+package com.yourssu.unscramble.util.extension
 
 import android.text.Html
 import android.text.Spanned
